Nailed the cause of the mismarking of some dstxs as unstructured.
The cause of the bug was the lines in dstx-text-mixin:

                  (set! f-cursor (cursor:replace
                                  f-cursor
                                  (send current-text
                                        decorate-new-dstx
                                        (struct:cursor-dstx f-cursor))))

This inadvertently took existing dstxs, removing them, and creating
new ones in a context that may try to annotate new dstxs as
unstructured.  The original purpose of the code is no longer valid
anyway, as it was meant to annotate sentinels when they were
automatically introduced.  Since we use decorate on fusions, which
cover them, this kludge is no longer necessary.